requirements.txt
azure-ai-translation-text==1.0.1
azure-core==1.34.0
certifi==2025.6.15
charset-normalizer==3.4.2
idna==3.10
isodate==0.7.2
python-dotenv==1.1.0
requests==2.32.4
six==1.17.0
typing_extensions==4.14.0
urllib3==2.5.0
